requirements.txt
streamlit-nightly
tiktoken

langchain_core 
langchain_community 
langchain_openai

PyPDF2
pdf2image

Authlib